rename ChefCLI::Telemtry to ChefCLI::Telemeter
This naming is more accurage - telemeter is the thing that captures
data, while telemetry is the external service it reports to.
