+++ components/chef-cli/lib/chef-cli/telemeter/patch.rb
@@ -1,4 +1,4 @@
-class ::Telemetry
+class Telemetry
   class Session
     # The telemetry session data is normally kept in .chef, which we don't have.
     def session_file
@@ -7,7 +7,7 @@ class ::Telemetry
   end
 
   def deliver(data = {})
-    if ChefCLI::Telemetry.instance.enabled?
+    if ChefCLI::Telemeter.instance.enabled?
       payload = event.prepare(data)
       client.await.fire(payload)
     end
