+++ components/chef-cli/spec/unit/action/base_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 require "chef-cli/action/base"
-require "chef-cli/telemetry"
+require "chef-cli/telemeter"
 require "chef-cli/target_host"
 
 RSpec.describe ChefCLI::Action::Base do
@@ -23,7 +23,7 @@ RSpec.describe ChefCLI::Action::Base do
 
   context "#run" do
     it "runs the underlying action, capturing timing via telemetry" do
-      expect(ChefCLI::Telemetry).to receive(:timed_action_capture).with(subject).and_yield
+      expect(ChefCLI::Telemeter).to receive(:timed_action_capture).with(subject).and_yield
       expect(action).to receive(:perform_action)
       action.run
     end
@@ -31,7 +31,7 @@ RSpec.describe ChefCLI::Action::Base do
     it "invokes an action handler when actions occur and a handler is provided" do
       @run_action = nil
       @args = nil
-      expect(ChefCLI::Telemetry).to receive(:timed_action_capture).with(subject).and_yield
+      expect(ChefCLI::Telemeter).to receive(:timed_action_capture).with(subject).and_yield
       expect(action).to receive(:perform_action) { action.notify(:test_success, "some arg", "some other arg") }
       action.run { |action, args| @run_action = action; @args = args }
       expect(@run_action).to eq :test_success
